[유치수] [NOBTS] terminal node request count incr.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-web/trunk@958 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/hippo/web/mapper/TerminalRequestCountMapper.java
@@ -28,7 +28,7 @@ public class TerminalRequestCountMapper implements RowMapper<List<TerminalReques
 			if (kv.getFamilyLength() == HBaseTables.TERMINAL_STATISTICS_CF_COUNTER.length) {
 				String from = TerminalSpanUtils.getApplicationNameFromRowKey(kv.getRow());
 				String to = TerminalSpanUtils.getApplicationNameFromColumnName(kv.getQualifier());
-				int requestCount = Bytes.toInt(kv.getValue());
+				long requestCount = Bytes.toLong(kv.getValue());
 				short serviceType = TerminalSpanUtils.getServiceTypeFromColumnName(kv.getQualifier());
 
 				TerminalRequest request = new TerminalRequest(from, to, serviceType, requestCount);
